requirements.txt
dj_database_url==0.5.0
Django==4.2.2
django_storages==1.10.1
djangorestframework==3.14.0
djangorestframework_api_key==2.3.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.7
openpyxl==3.1.2
Pillow==9.5.0
Pillow==10.2.0
python_dateutil==2.8.2